pyproject.toml
[project]
name = "nexor"
version = "0.2.0"
description = "Shared database and configuration helpers for the FDE services"
authors = [
    { name = "FDE Team", email = "support@riskary.de" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.27.0",
    "sqlmodel>=0.0.27",
    "sqlalchemy>=2.0.18",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.11.0",
    "tenauth @ git+https://github.com/thwolter/fde-tenauth.git@main",
    "loguru>=0.7.3",
    "opentelemetry-instrumentation-openai>=0.49.6",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-sdk>=1.38.0",
    "opentelemetry-exporter-otlp-proto-http>=1.38.0",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "uv"
update_changelog_on_bump = true
major_version_zero = true
[dependency-groups]
dev = [
    "commitizen>=4.10.0",
    "isort>=7.0.0",
    "pre-commit>=4.5.0",
    "pyrefly>=0.43.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.7",
    "ssort>=0.15.0",
    "typer>=0.20.0",
]


[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "single"


[tool.pytest.ini_options]
# Limit discovery & keep noise low
testpaths = ["tests"]
addopts = "-q -ra --tb=short --show-capture=no -p no:warnings"
asyncio_mode = "auto"
